requirements.txt
openai
python-dotenv
streamlit
pandas
json
pycaret.clustering
plotly
os
#pathlib